fixup! ISSUE-86 [Secret link] Implement DavClient - refactor CalendarURL
